Allow gRPC to be built without NIOSSL
Motivation:

In some situations NIOSSL isn't required when using gRPC. Since NIOSSL
brings a copy of BoringSSL with it it is quite a large depenency to
bring along.

Modifications:

- Wrap all code which uses NIOSSL in `#if canImport(NIOSSL)`
- This required shuffling around a few files, the code isn't
  meaningfully different otherwise

Result:

gRPC can be compiled without NIOSSL.
