+++ Tests/GRPCTests/ClientTLSTests.swift
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#if canImport(NIOSSL)
 import EchoImplementation
 import EchoModel
 import Foundation
@@ -210,3 +211,5 @@ private class AuthorityCheckingEcho: Echo_EchoProvider {
     preconditionFailure("Not implemented")
   }
 }
+
+#endif // canImport(NIOSSL)
