+++ Tests/GRPCTests/ConnectionPool/GRPCChannelPoolTests.swift
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#if canImport(NIOSSL)
 import EchoImplementation
 import EchoModel
 import GRPC
@@ -435,3 +436,5 @@ final class GRPCChannelPoolTests: GRPCTestCase {
     }
   }
 }
+
+#endif // canImport(NIOSSL)
