+++ Tests/GRPCTests/MutualTLSTests.swift
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#if canImport(NIOSSL)
 import EchoImplementation
 import EchoModel
 @testable import GRPC
@@ -267,3 +268,5 @@ class MutualTLSTests: GRPCTestCase {
     )
   }
 }
+
+#endif // canImport(NIOSSL)
