+++ Tests/GRPCTests/ZeroLengthWriteTests.swift
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#if canImport(NIOSSL)
 import Dispatch
 import EchoImplementation
 import EchoModel
@@ -261,3 +262,5 @@ final class ZeroLengthWriteTests: GRPCTestCase {
     #endif
   }
 }
+
+#endif // canImport(NIOSSL)
